Move the opentelemetry-jaeger-propagator module from to its own file
